+++ core-ng/src/main/java/core/framework/internal/web/HTTPServer.java
@@ -4,7 +4,6 @@ import core.framework.internal.async.ThreadPools;
 import core.framework.internal.log.LogManager;
 import core.framework.internal.web.site.SiteManager;
 import core.framework.internal.web.sse.ServerSentEventHandler;
-import core.framework.internal.web.websocket.WebSocketHandler;
 import core.framework.util.StopWatch;
 import core.framework.web.Interceptor;
 import io.undertow.Undertow;
@@ -36,22 +35,21 @@ public class HTTPServer {
 
     public final SiteManager siteManager = new SiteManager();
     public final HTTPHandlerContext handlerContext = new HTTPHandlerContext();
-    public final HTTPHandler handler;
+    public final HTTPHandler httpHandler;
     final ShutdownHandler shutdownHandler = new ShutdownHandler();
     private final Logger logger = LoggerFactory.getLogger(HTTPServer.class);
     private final ExecutorService worker = ThreadPools.virtualThreadExecutor("http-handler-");
 
-    public WebSocketHandler webSocketHandler;
     public ServerSentEventHandler sseHandler;
     private Undertow server;
 
     public HTTPServer(LogManager logManager) {
-        handler = new HTTPHandler(logManager, siteManager.sessionManager, siteManager.templateManager, handlerContext);
+        httpHandler = new HTTPHandler(logManager, siteManager.sessionManager, siteManager.templateManager, handlerContext);
     }
 
     public void start(HTTPServerConfig config) {
         var watch = new StopWatch();
-        handler.interceptors = config.interceptors.toArray(new Interceptor[0]);
+        httpHandler.interceptors = config.interceptors.toArray(new Interceptor[0]);
         HTTPHost httpHost = config.httpHost;
         HTTPHost httpsHost = config.httpsHost();
         try {
@@ -92,7 +90,7 @@ public class HTTPServer {
     }
 
     private HttpHandler handler(HTTPServerConfig config) {
-        HttpHandler handler = new HTTPIOHandler(this.handler, shutdownHandler, config.maxEntitySize, sseHandler, webSocketHandler);
+        HttpHandler handler = new HTTPIOHandler(httpHandler, shutdownHandler, config.maxEntitySize, sseHandler);
         if (config.gzip) {
             // only support gzip, deflate is less popular
             handler = new EncodingHandler(handler, new ContentEncodingRepository()
@@ -105,7 +103,6 @@ public class HTTPServer {
         if (server != null) {
             logger.info("shutting down http server");
             shutdownHandler.shutdown();
-            if (webSocketHandler != null) webSocketHandler.shutdown();
             if (sseHandler != null) sseHandler.shutdown();
         }
     }