+++ core-ng/src/test/java/core/framework/internal/web/HTTPIOHandlerTest.java
@@ -23,7 +23,7 @@ class HTTPIOHandlerTest {
 
     @BeforeEach
     void createHTTPIOHandler() {
-        handler = new HTTPIOHandler(null, null, 1000, null, null);
+        handler = new HTTPIOHandler(null, null, 1000, null);
     }
 
     @Test