+++ core-ng/src/test/java/core/framework/internal/web/sse/ServerSentEventWriterTest.java
@@ -5,20 +5,20 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class ServerSentEventBuilderTest {
-    private ServerSentEventBuilder<TestEvent> builder;
+class ServerSentEventWriterTest {
+    private ServerSentEventWriter<TestEvent> builder;
 
     @BeforeEach
     void createServerSentEventBuilder() {
-        builder = new ServerSentEventBuilder<>(TestEvent.class);
+        builder = new ServerSentEventWriter<>(TestEvent.class);
     }
 
     @Test
-    void build() {
-        assertThat(builder.build("001", "message"))
+    void message() {
+        assertThat(builder.message("001", "message"))
             .asString().isEqualTo("id: 001\ndata: message\n\n");
 
-        assertThat(builder.build(null, "message"))
+        assertThat(builder.message(null, "message"))
             .asString().isEqualTo("data: message\n\n");
     }
 }