+++ core-ng/src/main/java/core/framework/internal/kafka/MessageProducer.java
@@ -43,7 +43,6 @@ public class MessageProducer {
         producer.send(record, new KafkaCallback(record));
     }
 
-    @SuppressWarnings("deprecation")
     Producer<byte[], byte[]> createProducer(KafkaURI uri) {
         var watch = new StopWatch();
         try {
@@ -55,7 +54,7 @@ public class MessageProducer {
                 ProducerConfig.RECONNECT_BACKOFF_MAX_MS_CONFIG, 5_000L,                      // 5s
                 ProducerConfig.MAX_BLOCK_MS_CONFIG, 30_000L,                                 // 30s, metadata update timeout, shorter than default, to get exception sooner if kafka is not available
                 ProducerConfig.MAX_REQUEST_SIZE_CONFIG, maxRequestSize,
-                ProducerConfig.AUTO_INCLUDE_JMX_REPORTER_CONFIG, Boolean.FALSE,
+                ProducerConfig.METRIC_REPORTER_CLASSES_CONFIG, EmptyMetricsReporter.class.getName(),
                 ProducerConfig.ENABLE_METRICS_PUSH_CONFIG, Boolean.FALSE);
 
             var serializer = new ByteArraySerializer();