+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * sse: support @LimitRate on sse connect or class
   > use @LimitRate and http().limitRate().add() to configure rate limit
 * sse: log event count and size on sse:close action
+* ws: remove ws support
+  > websocket is not used anymore, use sse/ajax instead
 
 ### 9.1.7 (2/26/2025 - 3/6/2025)