+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * sse: log event count and size on sse:close action
 * ws: remove ws support
   > websocket is not used anymore, use sse/ajax instead
+* kafka: updated to 4.0.0
 
 ### 9.1.7 (2/26/2025 - 3/6/2025)